+++ GridironBulgaria.Web/Controllers/PhotoAlbumsController.cs
@@ -49,6 +49,11 @@
         [Authorize(Policy = "RequireAdminRole")]
         public async Task<IActionResult> Delete(int id)
         {
+            if (!this.ModelState.IsValid)
+            {
+                return NotFound();
+            }
+
             await this.photoAlbumsService.DeleteByIdAsync(id);
 
             return this.RedirectToAction(nameof(this.Index));
